+++ lib/src/controllers/picture_cropper_controller.dart
@@ -129,4 +129,84 @@ class PictureCropperController extends ChangeNotifier {
   void pictureEditorControllerDispose() {
     _cameraController?.dispose();
   }
+
+  /// Takes a Uint8List and returns the image extension type.
+  String getImageTypeFromBytes(Uint8List data) {
+    if (data[0] == 0x89 &&
+        data[1] == 0x50 &&
+        data[2] == 0x4E &&
+        data[3] == 0x47 &&
+        data[4] == 0x0D &&
+        data[5] == 0x0A &&
+        data[6] == 0x1A &&
+        data[7] == 0x0A) {
+      return 'png';
+    }
+
+    if (data[0] == 0xFF && data[1] == 0xD8 && data[2] == 0xFF) {
+      return 'jpg';
+    }
+
+    if (data[0] == 0x47 &&
+        data[1] == 0x49 &&
+        data[2] == 0x46 &&
+        (data[3] == 0x38 &&
+            (data[4] == 0x39 || data[4] == 0x37) &&
+            data[5] == 0x61)) {
+      return 'gif';
+    }
+
+    if (data[0] == 0x42 && data[1] == 0x4D) {
+      return 'bmp';
+    }
+
+    if (data[0] == 0x49 &&
+        data[1] == 0x49 &&
+        data[2] == 0x2A &&
+        data[3] == 0x00) {
+      return 'tiff';
+    }
+
+    if (data[0] == 0x4D &&
+        data[1] == 0x4D &&
+        data[2] == 0x00 &&
+        data[3] == 0x2A) {
+      return 'TIFF';
+    }
+
+    if (data[0] == 0x52 &&
+        data[1] == 0x49 &&
+        data[2] == 0x46 &&
+        data[3] == 0x46 &&
+        data[8] == 0x57 &&
+        data[9] == 0x45 &&
+        data[10] == 0x42 &&
+        data[11] == 0x50) {
+      return 'webp';
+    }
+
+    if (data[4] == 0x66 &&
+        data[5] == 0x74 &&
+        data[6] == 0x79 &&
+        data[7] == 0x70 &&
+        data[8] == 0x68 &&
+        data[9] == 0x65 &&
+        data[10] == 0x69 &&
+        data[11] == 0x63) {
+      return 'heif';
+    }
+
+    if (data[4] == 0x66 &&
+        data[5] == 0x74 &&
+        data[6] == 0x79 &&
+        data[7] == 0x70 &&
+        data[8] == 0x6D &&
+        data[9] == 0x69 &&
+        data[10] == 0x66 &&
+        data[11] == 0x31) {
+      return 'heic';
+    }
+
+    return 'unknown';
+  }
 }